Fix all build warnings in ASP3 common sources (zero warnings)
Address every -Wconversion, -Wunused-parameter, and related warning
flagged by Clang 18 with -Wall -Wextra -Wconversion on Cortex-M:

- sil.h: cast TOPPERS_SIL_REV_ENDIAN_UINT16 result to uint16_t (4 sites)
- arm_m.h: guard EXC_RETURN_PREFIX with #ifndef to avoid CMSIS redefinition
- core_kernel_impl.h/c: cast INT_IPM call sites to uint32_t (assembly
  macro constraint prevents changing the macro itself); use 1U in shift
  expressions; add (void)param stubs for unused parameters
- kernel/wait.h/c, task.c, task_manage.c, mutex.c: cast assignments to
  uint8_t bitfield members
- kernel/mempfix.c: cast ptrdiff_t pointer subtraction to size_t
- kernel/time_manage.c: cast int32_t adjtim to EVTTIM before addition
- kernel/startup.c: suppress unused variable warning
- library/log_output.c: cast char arithmetic result to uint_t
- syssvc/{syslog,serial,logtask,banner}.c: add (void)exinf to callbacks
- sample/sample1.c: add (void)exinf; remove unused cnt variable

+++ asp3/kernel/task.c
@@ -288,8 +288,8 @@ void
 make_dormant(TCB *p_tcb)
 {
 	p_tcb->tstat = TS_DORMANT;
-	p_tcb->bpriority = p_tcb->p_tinib->ipriority;
-	p_tcb->priority = p_tcb->p_tinib->ipriority;
+	p_tcb->bpriority = (uint8_t)p_tcb->p_tinib->ipriority;
+	p_tcb->priority = (uint8_t)p_tcb->p_tinib->ipriority;
 	p_tcb->wupque = false;
 	p_tcb->raster = false;
 	p_tcb->enater = true;
@@ -337,7 +337,7 @@ change_priority(TCB *p_tcb, uint_t newpri, bool_t mtxmode)
 	uint_t	oldpri;
 
 	oldpri = p_tcb->priority;
-	p_tcb->priority = newpri;
+	p_tcb->priority = (uint8_t)newpri;
 
 	if (TSTAT_RUNNABLE(p_tcb->tstat)) {
 		/*
