+++ asp3/kernel/wait.h
@@ -83,7 +83,7 @@ queue_insert_tpri(QUEUE *p_queue, TCB *p_tcb)
 Inline void
 make_wait(uint_t tstat, WINFO *p_winfo)
 {
-	p_runtsk->tstat = tstat;
+	p_runtsk->tstat = (uint8_t)tstat;
 	make_non_runnable(p_runtsk);
 	p_runtsk->p_winfo = p_winfo;
 	p_winfo->p_tmevtb = NULL;
